Convert all commands to objects'
