+++ src/main/java/com/command/Concatenate.java
@@ -1,0 +1,57 @@
+package com.command;
+
+import com.Constants;
+import com.utility.Util;
+import core.ZOSConnection;
+import org.apache.commons.io.IOUtils;
+import org.beryx.textio.TextTerminal;
+import utility.UtilIO;
+import zosfiles.ZosDsnDownload;
+import zosfiles.input.DownloadParams;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.StringWriter;
+import java.util.Arrays;
+
+public class Concatenate {
+
+    private final TextTerminal<?> terminal;
+    private final ZOSConnection connection;
+
+    public Concatenate(TextTerminal<?> terminal, ZOSConnection connection) {
+        this.terminal = terminal;
+        this.connection = connection;
+    }
+
+    public void cat(String dataSet, String param) {
+        final var dl = new ZosDsnDownload(connection);
+        final var dlParams = new DownloadParams.Builder().build();
+        InputStream inputStream;
+        try {
+            if (Util.isDataSet(param)) {
+                inputStream = dl.downloadDsn(String.format("%s", param), dlParams);
+            } else {
+                inputStream = dl.downloadDsn(String.format("%s(%s)", dataSet, param), dlParams);
+            }
+            display(inputStream);
+        } catch (Exception e) {
+            if (e.getMessage().contains("Connection refused")) {
+                terminal.printf(Constants.SEVERE_ERROR + "\n");
+                return;
+            }
+            Util.printError(terminal, e.getMessage());
+        }
+    }
+
+    private void display(InputStream inputStream) throws IOException {
+        if (inputStream != null) {
+            var writer = new StringWriter();
+            IOUtils.copy(inputStream, writer, UtilIO.UTF8);
+            String[] content = writer.toString().split("\\n");
+            Arrays.stream(content).forEach(terminal::println);
+        }
+        inputStream.close();
+    }
+
+}
